// libs/infra-status/src/api/node.docker.ts
// Generated by https://quicktype.io

export interface DockerNode {
  ID:             string;
  Version:        Version;
  CreatedAt:      string;
  UpdatedAt:      string;
  Spec:           Spec;
  Description:    Description;
  Status:         Status;
  ManagerStatus?: ManagerStatus;
}

export interface Description {
  Hostname:  string;
  Platform:  Platform;
  Resources: Resources;
  Engine:    Engine;
  TLSInfo:   TLSInfo;
}

export interface Engine {
  EngineVersion: string;
  Plugins:       Plugin[];
}

export interface Plugin {
  Type: Type;
  Name: string;
}

export enum Type {
  Log = "Log",
  Network = "Network",
  Volume = "Volume",
}

export interface Platform {
  Architecture: string;
  OS:           string;
}

export interface Resources {
  NanoCPUs:          number;
  MemoryBytes:       number;
  GenericResources?: GenericResource[];
}

export interface GenericResource {
  NamedResourceSpec: NamedResourceSpec;
}

export interface NamedResourceSpec {
  Kind:  string;
  Value: string;
}

export interface TLSInfo {
  TrustRoot:           string;
  CertIssuerSubject:   string;
  CertIssuerPublicKey: string;
}

export interface ManagerStatus {
  Leader:       boolean;
  Reachability: string;
  Addr:         string;
}

export interface Spec {
  Labels:       Labels;
  Role:         string;
  Availability: string;
}

export interface Labels {
}

export interface Status {
  State: string;
  Addr:  string;
}

export interface Version {
  Index: number;
}